day update

+++ 06-XCWeibo/Objc-UI/XCSwitch/XCSwitchDemoController.m
@@ -33,7 +33,7 @@
 }
 
 - (void)sliderView:(XMSSliderView *)view didSelectedAtIndex:(NSUInteger)index{
-    NSLog(@"%zd",index);
+    NSLog(@"%lu",(unsigned long)index);
 }
